+++ CHANGELOG.md
@@ -14,6 +14,9 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 ## [Unreleased]
 
+### Added
+- metadata parameter to IssuingPurchase resource
+
 ## [0.7.0] - 2023-05-11
 ### Added
 - payerId and endToEndId parameter to BrcodePreview resource